working docker, refactored folders, added qs
