+++ Dockerfile.worker
@@ -14,8 +14,3 @@ WORKDIR /app
 
 # Install dependencies
 RUN uv sync --locked
-
-# Expose the port FastAPI will run on
-EXPOSE 80
-
-CMD ["/app/.venv/bin/fastapi", "run", "dev", "--port", "80", "--host", "0.0.0.0"]
